fix name entry
The idea here is to use ScreenNameEntryTraditional's custom methods to
determine which players, if any, are entering high scores names, and
then pass those booleans off to a custom name entry screen
(ScreenNameEntryActual) which handles it from there.

This commit is large enough in size/scope that it really should be
multiple commits.