add custom failtype (fail song, but continue set)
If enabled, this allows players to fail songs but continue playing their set to completion.

There is a bit of hackery involved in that SM seems to believe that when all available players have failed, there are no stages remaining, yielding an empty MusicWheel.  I circumvented this with GAMESTATE:AddStageToPlayer() to reinstate stages as necessary.