use one global table instead of many env variables
This commit introduces ./Scripts/SL_Init.lua, which creates a single global table, SL, in an effort to pollute the global namespace less.  Any files that previously used setenv() or getenv() to pass data between screens now use the SL table.

This commit also starts to eradicate use of GAMESTATE:ApplyGameCommand() to set player modifiers in favor of kyzentun's new playeroptions interface.  As a result, player SpeedMod values are no longer stored as strings like "1.15x" or "C400" but as numbers so that they can be more easily passed into the playeroptions interface.

If you've been using this theme with a profile (local or memorycard), it will be necessary to delete your modifier .cfg files and start over.  Sorry about that.