overhaul ScreenNameEntryTraditional
My old NameEntry system was written prior to the engine's input callback system and was (under the hood) actually a PlayerOptions screen which allowed me to exploit repeating input.  It was an embarrassing hack and simply not necessary any longer.

This commit vastly overhauls NameEntry in Simply Love.

1. It is no longer necessary to have two distinct screens (Traditional + Actual).  Flags are now set appropriately from ScreenEvaluation if a player has earned the ability to enter a highscore name.

2. Input is handled using SM5's input callback system.

3. The AlphabetWheels now use Consensual's sick_wheel and require ~6x fewer Actors onscreen.  This will hopefully result in a more responsive NameEntry screen.

It is possible that some bugs are still lurking, but this seems reliable enough for now to commit!