bandaid fix for Issue #15
Player profile settings are currently read from .cfg files using the RageFile interface.  The RageFile:Read() method always returns Lua strings, which results in strings like "false" or "77.41"

In the case of Issue #15, the strings "true" and "false" would both evaluate as the boolean true; it was necessary to convert with a simple check.  Thus, this serves as a bandaid fix for Issue #15.

This problem could be circumvented, on the whole, by using .lua files to store player settings in their native datatypes.  That is a slightly bigger issue for another commit, however.