fix "purple banner" bug for nonexistent banners
This fixes issue #22 to the best of my knowledge and abilities.  As pointed out in pull request #23, the issue occurred when a song referenced a banner that did not actually exist.  The Sprite object would use the string path for lookup, find nothing, and yield a purple quad(?).

This commit fixes this by using the slightly more specialized Banner objects (in place of straight-up Sprite objects) which have a LoadFromCachedBanner method.  This circumvents the need to actually lookup whether the file exists or not using FILEMAN as proposed in pull request #23.