fix issue #24 by not loading Cached banners
Cached banners are typically blurry because Preferences.ini sets

BannerCache=LowResPreload

by default.  Users can, of course, set the value to

BannerCache=Full

but that's not necessarily feasible if a lot of songs are installed or available memory is anything but excessive(!).  Attempting to start SM5 with BannerCache set to "Full" resulted in OOM errors for me on an Arch box with "only" 4GB memory.

Anyway, most users won't even know BannerCache is a thing, and will have it set to the default of "LowResPreload."  While loading banners from cache for ScreenEval and NameEntry does read from disk less, it produces blurry banners.

This commit should address issues #24 (blurry banners) and #22 (missing banners) by using Banner:LoadFromSong() and Banner:LoadFromCourse() as appropriate.