load all available Vocalizations upon init
Previously, I had tried to load the the Lua vocalization tables as-needed upon reaching ScreenEvaluation.  This was kind of neat, but created problems if a player had an invalid Vocalization setting saved to profile.  ScreenEvaluation would attempt to load a non-existent default.lua and crash.

This commit forgoes the novelty of loading Lua as-needed and loads all available vocalization data into a table upon initializing the theme's ./Scripts/ directory.  This should resolve issue #19.

As usual, I've added lots of comments explaining how the code works in the hope that someone might learn from it.