fix most MiniMenus in Edit Mode
Some of the MiniMenus (most notably ScreenMiniMenuTimingDataInformation) were extended off the screen on the y-axis.  This commit fixes those oversights.

This also adds a few comments to metrics.ini explaining what different screens are for.

I also removed "Hide" as an available OptionRow from ScreenAttackMenu, as it doesn't really make sense to be there (the modifiers it contains won't work outside of this theme).