simplify SL-Colors.lua
This commit adds a new member to the global SL table, ActiveColorIndex, which stores the index to be used against the SL.Colors table.

This reduces the number of functions floating around in the global namespace.  Next up (eventually) is SL-PlayerOptions.lua.

This commit also includes a minor cleanup to ./BGAnimations/_shared background rainbow/default.lua

I had been experimenting with Def.ActorProxy{} actors in this file, but ultimately ended up not using them.  For future reference, ActorProxy inherits from Actor, so if you SetTarget() to, say, a Sprite actor, the ActorProxy won't be able use methods like customtexturerect on the ActorProxy.