clean up SL-PlayerOptions.lua
Previously, each custom OptionRow for ScreenPlayerOptions required its own function that was in the global Lua namespace.  This commit does away with each of those, replacing them with a single function, CustomOptionRow()

This commit also creates a reusable "generic" OptionRow metatable and a secondary table of OptionRow definitions that override individual aspects of the metatable as needed.

All in all, this continues my efforts to reduce the quantity of stray functions polluting the global namespace rather nicely.

(...and there are probably undiscovered bugs lurking somewhere in here.)