major cleanup for ScreenSelectMusic
This commit does away with many of the terrible coding conventions I utilized when this screen was originally written.  There should now be fewer magic numbers, more local variables, and greater reliance for sizing/positioning calculations to be does relative to other things on-screen.

There are *still* issues with the cursors when the StepsDisplayList has to handle both players with seven or more difficulties (Tachyon Alpha - Delta Max), but there always were.