add proper stream parser for MeasureCounter
This commit adds Lua that will parse .sm and .ssc files for consecutive measures of "stream" in configurable quantizations (8th, 12th, 16th, 24th, 32nd, etc).  This is used to enhance the existing MeasureCounter feature.

Previously, the MeasureCounter would only increment as it went, so players were left in the dark as to how many measures long the current stream might be.  With this commit, the stepdata can be parsed prior to loading ScreenGameplay, allowing players to see their current_measure out of total_measures.

There may still be bugs lurking, but this seems usable enough to push to the master branch at this time.

Many thanks to Jonathan Knepp for helping make this possible.  The SL_StreamParser.lua file that this commit adds is taken (with minimal modifications) from: https://github.com/JonathanKnepp/SM5StreamParser