handle negative StageTokens for not-EventMode
As the inline comment states:  This is somewhat hackish, but it serves to counteract Lua Hacks.

If ScreenGameplay was reloaded by a "gimmick" chart, then it is very possible that the Engine's concept of remaining stages will be incongruent with the Theme's.  Add stages back, Engine-side, if necessary.

This prevents us from returning to ScreenSelectMusic with negative stage tokens (engine-side) which manifests as an empty MusicWheel.