add HarshHotLifePenalty to SL-Static.lua
As pointed out by Mute:

"I think there is one more thing to tweak in the Stomperz life system, which is the amount of health the player loses on the first miss. I recall Matt discussing how stock ITG has it hardcoded that the first miss from full life docks you 10% of your lifebar; with stock ITG's life delta values, this wasn't very noticeable. But for Stomperz, this means that the first miss from full does 2.5x more damage than it's technically supposed to. This was unavoidable in the 3.95 port, and I *think* this is still happening with the current SM5 theme build, based on how much the surround lifebar seems to dip."

The engine's source code for the HarshHotLifePenalty preference seems to match the description above.  Turning it off (setting it to 0) in StomperZ mode and turning it on for all other modes should resolve the issue.

The LifeMeter should now have a consistent behavior whether it is full ("hot") or not in StomperZ mode.