use AFT for screenshot animation on ScreenEval
Previously, the screenshot animation worked by
	1. writing the screenshot to disk
	2. reading it from disk and loading it into a Sprite

This had performance issues when reading/writing to slow disks and especially USB stick profiles.

This commit changes the method so that the screenshot animation uses an ActorFrameTexture containing an ActorProxy that is proxy-ing the entire screen.  That AFT renders its child ActorProxy of the entire screen to texture and then uses Sprite's SetTexture() method with the AFT's new texture.

This circumvents the need to read a texture from a potentially slow disk.