save profile UserPrefs to .lua rather than .cfg
Writing a lot of little plaintext .cfg files was problematic in that everything was read back in as a string.

This commit replaces the .cfg code with code that write the contents of SL[pn].ActiveModifiers to a lua file and saves that lua file to the player's profile, if there is one.

This is somewhat experimental yet, and deserves further testing.