add Set Duration to ScreenGameOver
This commit adds the duration of the set, in seconds, to ScreenGameOver when a profile is in use.

It's far from ideal, as it simply sums the length of each song played, and does not take
	1. quitting out of songs early
	2. rate mods
into consideration.

This commit also moves strings used in ScreenGameOver to en.ini so they can be translated.