calculate set duration more accurately
We want to be able to display the duration spent playing the entire set for ScreenGameover.  We could call GAMESTATE:GetCurrentSong():MusicLengthSeconds(), store those, and sum them, but that wouldn't take rate mods and quitting/backing out of songs into consideration.

What we do here, instead, is calculate the time actually spent on ScreenGameplay when its OffCommand is called.