rewrite (again) of SSM's SortMenu
This commit makes use of SCREENMAN:set_input_redirected() for ScreenSelectMusic's SortMenu overlay.  This method was added in SM5.0.11 (apparently), but I just learned of it now.  It's really the optimal solution for redirecting hardcoded-by-the-engine to Lua.

Thus, with this commit, it is no longer necessary to render SSM to texture and set that texture to the background of a new ScreenSortList just for the sake of input!  Nice.