scores on EvalSummary should always start white
This commit should fix a bug in which passing scores on ScreenEvaluationSummary were sometimes being colored red, as though they were failing scores.  The bug occurred when scrolling back and forth through the multiple pages of scores that often occur on ScreenEvaluationSummary.

The fix: each score on ScreenEvalSummary should always have a diffuse(Color.White) applied when navigating through pages of scores.

Immediately after applying that diffuse, check if the grade was a fail and, if so, apply a diffuse(Color.Red).