Pacemaker shouldn't block centered notefield
As pointed out in issue #46, the Pacemaker graph (contributed by Z-i-V user JackG) would overlap with the notefield when the preference Center1Player was set to true.  This commit should resolve that issue, and performs some minor cleanup to the Pacemaker code (no need to use MESSAGEMAN when a simple queuecommand() will suffice).