fix bug when changing GameModes from SSM
There was a longstanding bug involving the DecentsWayOffs modifier that occurred when switching from Casual -> Competitive, or from Casual -> StomperZ.

The modifier "DecentsWayOffs" should be "off" by default in Casual GameMode, and "on" by default in all others.  The bug was that if you entered ScreenSelectMusic in Casual mode (lets say by accident) and then immediately switched to Competitive mode (which is what you had really wanted), the "DecentsWayOffs" modifier stayed set to "off."

This commit should resolve that issue.