have SubtractiveScoring accommodate ECFA/StomperZ
This commit slightly tweaks the SubtractiveScoring logic so that it accommodates the FA+ timing windows used in ECFA and StomperZ GameModes.

There seems to be some lurking bug causing rolls to count twice, which throws off subtractive scoring, but this can be committed now as it is still better than it was.