temp fixes for misc Gameplay element positioning
ColumnFlashOnMiss, MeasureCounter, and SubtractiveScoring relied on the now-deprecated Style:GetWidth().  This commit replaces those calls with a kind of kludgy function I apparently had in SL-Helpers.lua, GetNotefieldWidth().  This seems to work for now, but will likely break under more exotic conditions/styles/etc.

MeasureCounter has been moved from Gameplay's underlay to overlay so that it maintains its draw on top of the notefields.

Kyzentun is suggesting that ColumnFlashOnMiss should be part of NoteColumn layers so that "you don't need special positioning logic to figure out where the field actually is, and it'll stay attached when mods move the field or column."  I suppose I'll have to look into this at some point.