add CourseContentsList for CourseMode
Previously, ScreenSelectMusic was more or less identical for both regular gameplay and CourseMode.  This commit adds a CourseContentsList to ScreenSelectMusic when in CourseMode.  It takes the place of the grid of difficulty blocks that is normally there.

What is a CourseContentsList, you ask?  It's a list of each song and the meter of the steps from that song that you'll be playing during the course.

I'm not completely satisfied with this implementation yet, but it's a good start.

TODO: See if I can get the CourseContentsList to scroll "back up" once it reaches the end of the list.