add 4th Pane to ScreenEvaluation
This commit adds an experimental-for-now 4th Pane to ScreenEvaluation: a judgment density histogram.  I'm attempting to address the problem where players can't tell if they are consistently stepping early or late.

As of now, the histogram supports Competitive, ECFA, and StomperZ, as adapts to Decents and WayOffs being disabled.

This feature is considered experimental for now, as it hasn't been extensively tested.  There are probably conditions that will cause the histogram to misbehave or outright break.  Maybe it will perform badly on low-end PCs (if this is the case, I'll render the whole pane4 to an ActorFrameTexture).  Still, I want to get this out there for people to test and provide feedback.