tidy UI of offset density histogram
This commit tidies the UI of ScreenEvaluation's Pane4 somewhat, adding dark quads at the top and bottom, removing the mode_offset line, etc.  It's all still subject to change, but I wanted to commit what I have for now and get it on GitHub so others could try it out.

I've also added lots of comments to ./Pane4/default.lua for those those who are curious and like to poke around in my code.