bandaid fix for ChartParser
This commit adds a bandaid fix for ./Script/ChartParser.lua which previously would choke on charts from the "You're Streaming Again" pack.  This caused crashes for some users.

In SM/SSC files, one chart is separated from another by the presence of a semicolon.  In most files, this semicolon gets its own line, after the last row of the last measure like this:

0100
1000
0010
0001
;

... but it seems that that is not mandatory.  SM files from "You're Streaming Again" would include the semicolon on same line as the last row of the last measure, like this:

0100
1000
0010
0001;

Thus, this commit patches the ChartParser to account for this specific possibility.

I consider this a bandaid fix because it is very possible that some other charts exist that do other things that the SM5 engine accepts that will cause the pattern matching that this parser uses to choke.

teejusb has suggested that a proper fix would involve porting the engine's MSD format parser to Lua.  See: https://github.com/stepmania/stepmania/blob/master/src/MsdFile.cpp#L32-L143

That remains an open task for another day.