Highscores Character Limit increased from 4 to 10
I'm increasing the character limit on highscore names because there is
no reason why we can't fit 10 characters, especially now that we all
play in widescreen. If you're worried about some less-than-mature person
using a name like "COCKSLUT" then get over yourself; you're as immature
for being offended by a gamertag as they are immature for thinking that
it was funny in the first place.