Display song folder on the music selection screen
The song selection screen now shows either folder or group tags for the
selected song. You can toggle between a verbose "full" folder name and
just the group name in Simply Love Options.

The actual name of the folder where a simfile is located will show up
when folder mode is turned on. When you turn the option off, the music
select screen will show the selected song's group instead. The idea here
is that folders sometimes have useful information in them. My particular
usecase is that when you have multiple charts for the same song, knowing
which group a particular selection is from helps give context to which
song you are selecting. ("I want to play the origional Welcome To
Rainbow, not mute's "/aweful/" version")