NameEntry's starting char when using Profile is OK
This changes the default behavior of the AlphabetWheels on ScreenNameEntryTraditional.  If a player is using a profile, and the profile's HighScoreName isn't empty, the starting character is now the "OK" checkmark.

If a profile isn't found, or if the profile's HighScoreName is empty, then the starting character is still "A".