add judgment scatter plot to ScreenEvaluation
This adds a(n experimental for now) judgment scatter plot to ScreenEvaluation.  People have been requesting this.

Early judgments are represented as dots "below the middle."
Late judgments are represented as dots "above the middle."

This does not appear on ScreenEvaluation for CourseMode.  I haven't yet tested what happens when ScreenGameplay is paused.