add option for StepStatistics during gameplay
I wrote this quite a while ago, thinking I'd add a note density histogram and a judgment scatterplot to it, too, but those remain open tasks.

This commit adds a pane for single-player gameplay to show how many of each judgment the player has earned so far.  Due to it taking up half the screen, this feature is not available for 2-player/versus or double gameplay,  and is not available when "Center1Player" is enabled in Preferences.ini.

There are probably bugs lurking, but I've been sitting on this feature for nearly two months now, so I might as well push and get feedback from users.