fix Color change bug after visiting options
Code added by RainbowMode forced the SL.Global.ActiveColorIndex to be 3 as I had deemed that color to produce the most legible text against the ever-changing background colors.  This was enforced on ScreenOptionsService's BeginCommand.

The problem was that players who didn't use RainbowMode would still have their ActiveColorIndex value forcibly changed to 3 after visiting ScreenOptionsService, which would be reflected in text changing color, but not the background shapes.

This commit should fix issue #64 by adding a conditional check.  When visiting ScreenOptionsService, if RainbowMode, forcibly change the ActiveColorIndex to 3, if not change it to be whatever was last saved to ThemePrefs.