hide SSM fallback banner when not needed
Mute reported that, in 4:3 aspect ratio, the fallback banners were peeking out from behind the FadingBanner ActorProxy.

He provided some new zoom and positioning numbers that resolved that issue, but the zoom factor he provided (0.745) seemed to cause some strange artifact issues around the top edge of the FadingBanner ActorProxy...  Or, at least, it did on my Mac.

This led me to try a 0.75 zoom factor, which cleaned up the artifacts but causes about 2-3 pixels of the left side of the banner to be cut off.  Hmm.  I ... think this is probably preferable, but we'll see what people think.

I also added some code to just straight up hide the fallback banners when the current song or course has a banner.