attempt to fix NPS histogram clipping issues
Mute demonstrated that the right end of the NPS histogram was getting clipped off in certain circumstances.  This was due to improper horizontal scaling combined with the histogram AMV being a mask for the yellow gradient texture that was not sized appropriately.

This also adds some comments to the ChartParser script.