cleanup SystemLayer overlay
This commit is for me and the few others who debug in Simply Love using SCREENMAN:SystemMessage(thing) or the shorthand SM(thing) function.

When doing rapid debugging with SM() to print the value of variables to the screen, there was previously a Quad that was hardcoded to be 30px in height behind the text.  If the text was taller than 30px, it spilled out, over the Quad.  This was (mildly) annoying to me, so I fixed it to re-zoom the Quad's to accommodate the height of the SystemMessage text.

This also removes a large if-else chain that was (for once) better handled with a few additional lines in Metrics.ini.