Eval's ScatterPlot shouldn't alter JudgmentColors
The ScatterPlot adds an alpha of 0.666 to the JudgmentColors so that multiple datapoints (quadrilaterals) drawn in the same relative space will appear as less opaque (hopefully).  Unfortunately, my previous code was permanently altering the JudgmentColors table, so that everything that relied on them after the first ScreenEval would get colors with the 0.666 alpha value still intact.

This commit fixes that by using temp r, g, and b variables when defining the datapoints in the Scatterplot.