account for MusicRate on ScreenSelectMusic
This commit does a few things:

1. The BPM display on ScreenSelectMusic now accounts for MusicRate.  Previously, it always assumed the MusicRate was 1.  Now, if the BPM of a song is 100, and the MusicRate is 1.1, the BPM will display as 110.

2. Adds a Quad+MusicRate text to the bottom of the banner on ScreenSelectMusic if MusicRate is anything but 1.  This is visually similar to what happens on ScreenEvaluation.

3. Rewrote some of the code inside my GetDisplayBPMs() function to make this possible.

4. Rewrote some of the code inside ScreenPlayerOptions overlay.lua to fix what that^ broke.  In general, the code is now more modular with far fewer repeated "magic numbers."  There are now variables for the increments and upper/lower bounds to use when changing SpeedMod.

I also discovered that there were previously some rounding errors occurring within the help text on ScreenPlayerOptions in displaying what the BPM of the song was when MusicRate was anything but 1.  Oops.  Thankfully it didn't impact the SpeedMods themselves, just the text that displayed the song's current BPM.  This should be fixed now.

- - - - -
I tested this in normal gameplay and CourseMode, with one player and two, but it's likely I missed something and introduced new bugs here.