fix oversight when calculating CourseMode BPMs
It seems that calling GetSong() on a CourseEntry object from a randomly generated Course will return nil.  This commit circumvents the Lua error(s) that result in this scenario.

There may be a way to still calculate the BPM range for a randomly generate Course, but that will have to wait for a future commit.