more MusicRate fixes
CourseMode will attempt to briefly display the background art for the next/upcoming song just before that song starts.  While testing courses, I found that a high MusicRate could result in a scenario in which the background art was still being displayed when the arrows had already started.

This commit decreases the duration which the background is shown, and also checks to see if the MusicRate is > 1.  If it is, the hurrytweening() command is applied to the background's tweening to speed it up further yet.