fix NameEntry for multiple Courses
Some code in ScreenEvaluationSummary and ScreenNameEntryTraditional was previously written under the erroneous presupposition that it is only ever possible to play a single Course per game session.  In EventMode it is, in fact, possible to play multiple courses consecutively.

Some of this was already addressed in commit cb420e32a65a04e012576d4f4895deeb8ad77cd8.

This commit fixes ScreenNameEntryTraditional to correctly display HighScore information when multiple Courses have been played.

While testing this, I noticed that new HighScore rows don't "light up" in EventMode the way they do in not-EventMode.  This seems to be because calling GetMachineHighScoreIndex() on a PlayerStageStats object in EventMode always returns -1, which broke some of my code.

So, this commit also checks to see if EventMode is on, and attempts to "light up" the new HighScore row(s) via a slightly more roundabout check if necessary.