reduce number of points drawn in NPS histogram
"So I heard you like streams." AKA

"Reduce the number of datapoints being drawn in your density histogram by a factor of 18000 with this one weird trick." AKA

"I don't think dance games were ever intended to be like this."

---

This commit (typically) reduces the number of points in the verts table used to draw the note-per-second histogram that is part of the optional "Step Statistics" for Gameplay.

Some charts will see drastic reductions in the number of points.  The "24 Hours of 100bpm Stream" chart is reduced from 71998 points to 4.

"Last Trip" is reduced from 3148 to 650.

This should typically be a fairly modest optimization so long as it didn't break anything or (visually) cause any useful information to be lost.