fix errors regarding HideScore
If both players were joined and if one of those players opted enable HideScore, this would result in Lua errors within the WhoIsCurrentlyWinning.lua file, as it expected both players' score actors to be present.

This problem was pointed out and fixed in GitHub pull request #76 by GitHub user Kauhsa.  I used this as an opportunity/excuse to clean up some related files that were repeating the same actor lookups for every single JudgmentMessageCommand and didn't need to be.