add drop-shadow to subtractive scoring text
This commit adds a drop-shadow to the subtractive scoring BitmapText actor so that it is legible for players who aren't using a ScreenFilter but are using SubtractiveScoring.

Additionally, this commit switches SubtractiveScoring.lua from spaces to tabs to bring it in line with the rest of the project.

This commit effectively cherry-picks commit b4e003306ec249931a13bb26a33cec0797a50c9c out of the larger pull request it was part of.  Thanks go to GitHub user @andrewipark for the work there.