add optional MusicWheelItem letter grades
This feature was requested a number of times over the years by a number of users.  I never felt like working on it before, but GitHub user dougshell was kind enough to do the work to make it happen and write up a description of it for reddit.  Neat. :)

This slightly cleans up the code provided by dougshell on reddit and I spent some time (more than I want to admit because I am a photoshop noob) making a new spritesheet.  The sizing of the stars and letter grades should somewhat more closely match their ScreenEvaluation counterparts.

This is a Simply Love ThemePref and it disabled by default.  You can enable it under Simply Love Options in your operator menu.

This should resolve GitHub issue #71.