small performance improvement for StepStatistics
This attempts to address issue #89 on GitHub by potentially reducing the number of vertices drawn for StepStatistic's lifeline.

If Gameplay LifeMeter has been sampled at least twice already, and the current LifeMeter:GetLife() value hasn't changed any since the previous two samples, we now simply extend the previous vertex in the lifeline further to the right, rather than *always* pushing more vertices to the table.

As noted in the discussion on GitHub for issue #89, this won't fix every case, and truly nothing beats having a dedicated video card, but this should generally help performance a little.