Fixed hanging due to step descriptors having to finish their current loop
I couldn't figure out why using quecommand loops wouldn't stop queing
commands even after a song was selected. Turns out I needed to have an
offcommand that passed stoptweening(). In the past I was told that
stoptweening wouldn't work with this particular type of actor, but
whatever, it works as intended now. This behaviour is slightly different
from my experience in oITG where I had a similar marquee of actors
flashing on and off whereby after a song was selected, the looping
stopped and ScreenSelectMusic out played instead. Whatever the case may
be, I've fixed this annoying dead-end of a bug.