Song Folder/Group Now Displayed On ScreenSelectMusic/ScreenEvaluation
"Yo dude, what pack is this song from?"

Basically the question everyone asks whenever scores are posted. You can
choose between song folder and group on ScreenSelectMusic, but not on
ScreenEvaluation. The default is to show the song group; although that
option is configurable and found in the Operator Menu under Simply Love
Options (below Rainbow mode).

-ScreenEvaluation  will only display group name because I don't think
it's important to have redundant information (since the song folder will
almost always have the name of the song in it, which is displayed
directly below the group information). If there was an interest, it's
super easy to make this configurable in Simply Love options.

-Folder displays the full folder name of the song selected. Example:
"Eurobeat Is Fantastic/Deja Vu - [Arvin]" Both the name of the pack
folder, and the folder the song is found in are listed. This is great
for looking for specific songfiles, and especially since song folder
names usually have "useful" unformation in them.

-Group displays only the pack that the selected song is in. This is good
for arcade owners that are concerned about potential "questionable
language" in a songfolder name. I'm pretty sure arcade owners know
already not to put questionable content in their pack folder names. This
is also very useful if you're sorting the songwheel by anything other
than group.