use _wendy monospace for EventMode timer in header
The digits in the EventMode timer in SelectMusic's and Evaluation's header shouldn't bounce around any longer.

This required me to add a colon bitmap to "_wendy monospace" using Photoshop and update the ini.  I'm not the most adept with Photoshop, but I don't think I broke anything else that relies on this font.