remove unused code for Combo/Misses strings
Player combo.lua returned an ActorFrame containing two BitmapTexts – one for the combo digits, one for the label of "Combo" or "Misses".

Through a programming oversight I never discovered until now, the label would never be displayed.  And, you know, I've grown to like it that way, so this commit removes that code as it is basically just clutter.