fix input handling bug with SelectProfile
Previously, there were some conflicts between ScreenSelectProfile (which I'd basically nabbed out of _fallback) and Simply Love's AutoStyle ThemePref.  It was possible to break the theme if AutoStyle was set to "single" or "double" and a player joined both players on ScreenSelectProfile.

This commit should resolve that bug, and adds a few paltry comments explaining what I did.