support custom group list for SSMCasual
Conscientious public machine operators can now manually specify exactly which packs they want to show up in Casual Mode's ScreenSelectMusicCasual.  Packs titles can be added to ./Other/CasualMode-Groups.txt

If the file is absent or empty or contains only invalid pack names, the full list of song groups will be used as a fallback.

For now, I am including a CasualMode-Groups.txt file with some common packs that are good for beginners already filled in.  Depending on feedback, I intend to either expand this list or remove it altogether prior to the next release.

A separate readme.md explaining how to use the file in greater detail will follow in a future commit.