detangle credits from sides joined in SelectStyle
There was previously a bug with the (new) ScreenSelectStyle where it was possible to attempt to enter double with both players joined.  Double wants only one player joined, so this threw an error and the UI would get stuck.

That's fixed now, but I'm not convinced that:
1. there aren't additional bugs lurking
2. my current input handling is ideal

The old ScreenSelectStyle would invalidate single as a choice if both players were joined.  My new implementation of this screen doesn't do that.  Should it?  I think I'm going to need feedback from people using this on arcade hardware...