prefer player that presses Start over MPN
Previously, if both players were joined, but the player decided to choose single or double, I was always using the MasterPlayerNumber and unjoining the other side.  I think it makes more sense to prefer whichever side the start button was just pressed on, and unjoin the other side.

For double this doesn't really make any difference, because in double, only one player is joined, but that one player can use MenuButtons from either side to control the game.

This results in some slightly different behavior for single, however.  It's now possible to, say,

press P1 start on ScreenTitleJoin to make P1 the MPN
press P2 start on ScreenSelectStyle to join P2
    (both players are now joined)
press P2 start on ScreenSelectStyle to choose single
    (this previously wasn't possible)
play a game of single as P2

While the behavior is new/different, I think it's an improvement.  Maybe?