add audio/visual feedback to SelectProfile
ScreenSelectProfile prevents two players from both choosing the same local profile, but it lacked feedback to indicate that it was doing this.  Previously, if both players had been trying to choose the same profile, ScreenSelectProfile would play the confirm sound and then just sit there.

Now, I have the theme play the  "Common invalid" sound effect and gently shake the scroller frame belonging to the player that just pressed start.  Hopefully this is a little more clear.  I could add more feedback (text), but I'm not sure its necessary.

If both players have chosen unique profiles, the screen will now wait 0.5 seconds to give the scroller frames time to tween out before moving onto the next screen.  It was more abrupt before.

Finally, this fixes a bug with how input was handled on this screen if the ThemePref for autostyle was set to "single" or "double".  It's now possible to proceed past this screen again. :)